requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langgraph
openai
transformers
langgraph-supervisor
wikipedia
wolframalpha
langmem
greenlet
asyncpg
langchain_huggingface
sentence-transformers
pyowm
geopy
langchain-community
langchain-tavily
langgraph-api
PyGithub
langchain-openai
gtts
timezonefinder
geocoder
psycopg2
requests
aiogram
aiohttp
colorama
colorlog
pydantic
pyttsx3
pytz
pyyaml
regex
tiktoken
huggingface-hub
fsspec
typing-extensions
tqdm 
